requirements/base.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.28,<0.29
django-filter>=24.3,<25.0
django-cors-headers>=4.6,<5.0
django-environ>=0.11,<0.12
psycopg[binary]>=3.2,<4.0
celery>=5.4,<6.0
flower>=2.0,<3.0
redis>=5.2,<6.0
gunicorn>=23.0,<24.0
whitenoise>=6.7,<7.0
reportlab>=4.2,<5.0
pyotp>=2.9,<3.0
